refactor auto_detection and file_io into separate files
