+++ lib/smarter_csv.rb
@@ -3,7 +3,9 @@
 require "core_ext/hash"
 
 require "smarter_csv/version"
+require "smarter_csv/file_io"
 require "smarter_csv/options_processing"
+require "smarter_csv/auto_detection"
 require "smarter_csv/variables"
 require "smarter_csv/headers"
 require "smarter_csv/parse"
@@ -18,7 +20,7 @@ when 'ruby'
       require_relative "smarter_csv/smarter_csv"
       # :nocov:
     end
-  rescue Exception
+  rescue Exception # rubocop:disable Lint/RescueException
     #  require_relative 'smarter_csv/smarter_csv'
   end
 # :nocov:
